Implementation of CoCoSo Method

Implementation of CoCoSo Method

cocoso(example_cocoso, parameter_alpha = 0.5)

Arguments

md_data

Must be a data frame with information about Criteria, Alternatives and Optimal Value. See example_cocoso.

parameter_alpha

Parameter for third aggregation strategies.

Value

Final ranking of all alternatives from CoCoSo method. In addition, we compute and present all three aggregation strategies.

Examples

cocoso(example_cocoso)
#>     Alternatives Order        k       ki_a     ki_b      ki_c        Si
#> 4  Alternative_1     5 2.041313 0.13062317 3.245282 0.7242238 0.4299734
#> 5  Alternative_2     2 2.787990 0.17546935 4.473498 0.9728678 0.6082495
#> 6  Alternative_3     1 2.882350 0.18036300 4.639646 1.0000000 0.6363392
#> 7  Alternative_4     4 2.416046 0.16332196 3.720904 0.9055181 0.4470733
#> 8  Alternative_5     7 1.298692 0.08779670 2.000000 0.4867778 0.2402679
#> 9  Alternative_6     6 1.443143 0.09737208 2.225040 0.5398672 0.2683310
#> 10 Alternative_7     3 2.519094 0.16505373 3.951291 0.9151197 0.5031435
#>          Pi
#> 4  3.291361
#> 5  4.390712
#> 6  4.502038
#> 7  4.205820
#> 8  2.260980
#> 9  2.505710
#> 10 4.199086